pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "image-preference-modelling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.8",
    "gradio>=5.49.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.scripts]
gradio-cockpit = "image_preference_modelling.gradio_app:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/image_preference_modelling"]
